Add support for dynamic Forge build numbers
Specify the FORGE_BUILD environment variable to build for a specific
build of Forge, rather than the static value in the pom (preserved
for now for compatibility with existing ForgeAPI setup).

Intended for dynamically publishing new ForgeAPIs as new builds become
available. Currently only the build number can be specified, since the
major/minor/revision numbers are in the <version> tag of the pom.xml.